notifications send order fixed

+++ assets/templates/index/notification_template.js
@@ -7,10 +7,10 @@ import { $, notification } from "../../../js/utilities.js";
 bindInlineFunctions(getContext());
 
 export function notificationTemplate(notification) {
-  switch (notification.type) {
-    case "friendshipRequest": {
-      const { userName, userId, id } = notification;
-      return /*html*/ `
+	switch (notification.type) {
+		case "friendshipRequest": {
+			const { userName, userId, id } = notification;
+			return /*html*/ `
                 <div class="notification radius-15 row" data-id="${id}">
                     <div>
                         <span data-lang="user-has-sent-friendship-request, {ownerName: '${userName}'}"></span>
@@ -21,16 +21,16 @@ export function notificationTemplate(notification) {
                     </div>
                 </div>
             `;
-    }
-    case "boardInvite": {
-      const { ownerName, boardName, boardId, id } = notification;
-      return /*html*/ `
+		}
+		case "boardInvite": {
+			const { ownerName, boardName, boardId, id } = notification;
+			return /*html*/ `
                 <div class="notification radius-15 row" data-id="${id}">
                     <div>
                         <span data-lang="user-invited-you-to-board, {ownerName: '${ownerName}', boardName: '${boardName.replaceAll(
-        "-",
-        " "
-      )}'}"></span>
+				"-",
+				" "
+			)}'}"></span>
                     </div>
                     <div class="btn-container gap-10">
                         <button data-lang="decline" class="btn btn-secondary txt-small txt-600" onclick="removeNotification('${id}')">Decline</button>
@@ -38,10 +38,10 @@ export function notificationTemplate(notification) {
                     </div>
                 </div>
             `;
-    }
-    case "assignTo": {
-      const { userName, boardName, taskName, id } = notification;
-      return /*html*/ `
+		}
+		case "assignTo": {
+			const { userName, boardName, taskName, id } = notification;
+			return /*html*/ `
                 <div class="notification radius-15 column gap-10" data-id="${id}">
                     <div>
                         <span data-lang="assigned-you-to-task, {ownerName: '${userName}', taskName: '${taskName}', boardName: '${boardName}'}"></span>
@@ -51,47 +51,46 @@ export function notificationTemplate(notification) {
                     </div>
                 </div>
             `;
-    }
-    default:
-      return;
-  }
+		}
+		default:
+			return;
+	}
 }
 
 export async function acceptBoardInvite(boardId, boardName, notificationId) {
-  const USER = STORAGE.currentUser;
-  const BOARD = new Board(STORAGE.data.boards[boardId]);
-  await removeNotification(notificationId);
-  if (!(boardId in STORAGE.data.boards)) return notification(`board-nonexistent, {boardName: '${boardName}'}`);
-  USER.boards.push(boardId);
-  BOARD.collaborators.push(USER.id);
-  await Promise.all([USER.update(), BOARD.update()]);
-  await notification(`board-joined, {boardName: '${boardName}'}`);
-  location.reload();
+	const USER = STORAGE.currentUser;
+	const BOARD = new Board(STORAGE.data.boards[boardId]);
+	await removeNotification(notificationId);
+	if (!(boardId in STORAGE.data.boards))
+		return notification(`board-nonexistent, {boardName: '${boardName}'}`);
+	USER.boards.push(boardId);
+	BOARD.collaborators.push(USER.id);
+	await Promise.all([USER.update(), BOARD.update()]);
+	await notification(`board-joined, {boardName: '${boardName}'}`);
+	goTo('index/board/board');
 }
 
 export async function removeNotification(notificationId) {
-  const USER = STORAGE.currentUser;
-  delete USER.notifications[notificationId];
-  await USER.update();
-  $(`.notification[data-id="${notificationId}"]`).remove();
-  checkNotifications();
+	const USER = STORAGE.currentUser;
+	delete USER.notifications[notificationId];
+	await USER.update();
+	$(`.notification[data-id="${notificationId}"]`).remove();
+	checkNotifications();
 }
 
 export async function removeFriendshipRequest(id, userId) {
-  await STORAGE.delete(
-    `users/${userId}/pendingFriendshipRequests/${STORAGE.currentUserId()}`
-  );
-  return removeNotification(id);
+	await STORAGE.delete(`users/${userId}/pendingFriendshipRequests/${STORAGE.currentUserId()}`);
+	return removeNotification(id);
 }
 
 export async function acceptFriendshipRequest(id, userId, name) {
-  const user = STORAGE.currentUser;
-  const contact = STORAGE.allUsers[userId];
-  user.contacts.push(userId);
-  contact.contacts.push(STORAGE.currentUserId());
+	const user = STORAGE.currentUser;
+	const contact = STORAGE.allUsers[userId];
+	user.contacts.push(userId);
+	contact.contacts.push(STORAGE.currentUserId());
 
-  await Promise.all([user.update(), contact.update()]);
-  await removeFriendshipRequest(id, userId);
-  await notification(`friend-added, {name: '${name}'}`);
-  location.reload();
+	await Promise.all([user.update(), contact.update()]);
+	await removeFriendshipRequest(id, userId);
+	await notification(`friend-added, {name: '${name}'}`);
+	goTo('index/contacts/contacts');
 }
